refactor(cli): move list command deps into a dedicated module
